* commit '0191f020928b39c50c53e7f29be22ef9e3d5e5ac':
  BG-3970: avg bugfix AbgExpr
  BG-3970: avg bugfix AbgExpr

+++ yupana-core/src/main/scala/org/yupana/core/sql/FunctionRegistry.scala
@@ -92,8 +92,8 @@ object FunctionRegistry {
     ),
     uNum(
       "avg",
-      new Bind2[Expression, Numeric, Expression[Double]] {
-        override def apply[T](e: Expression[T], n: Numeric[T]): Expression[Double] = AvgExpr(e)(n)
+      new Bind2[Expression, Numeric, Expression[BigDecimal]] {
+        override def apply[T](e: Expression[T], n: Numeric[T]): Expression[BigDecimal] = AvgExpr(e)(n)
       }
     ),
     uTyped("year", TruncYearExpr),
